--- library/src/main/java/tech/summerly/streamcache/connection/NetUtils.kt
@@ -1,6 +1,0 @@
-package tech.summerly.streamcache.connection
-
-
-internal typealias HttpHeader = MutableMap<String, String>
-
-
